Add SystemProgram node for quick instructions
The instruction creation is tedious and multiple phantom signers was
impossible. This commit reworks signers by having phantom controllers in
the signers array and adding a sign at method that signs at a specific
index. This allows a more generic transaction wich in turn allows for an
easier instruction creation.