Add method to confirm transactions
When sending a transaction, the result is a transaction signature. There
is no easy way to find out if the transaction was confirmed or finalized
by the blockchain. This commit adds two new methods that performs a check
with the RPC endpoint if the transaction has become confirmed or
finilized.