Implement wallet adapter node
More wallet types must be supported. This commit enables more wallets to
be added to a new wallet adapter node. The old PhantomController is
replaced by WalletAdapter which can be configured to work with different
kinds of wallets. It does so by storing an enum preperty that can be set
to change wallet type. This commit also implements Solflare and Backpack.