Implement CandyMachine node
The instructions for minting from a candymachine is not  implemented. This
commit implements the instructions for initializing a candy machine and
minting from the machines without candy guard. It adds a new node for
candy machine program and static instruction creation methods. A pubkey
method for creating PDA with bytes is also implemented to be able to
create the instruction keys needed.