Add automatic tests
There are no way to know if changes are breaking functionality. This makes
maintainence crucial for new bugs on new commits. Automatic tests are
added to avoid this by creating a github action that runs all example
projects in the example directory. In this commit thex are also modified
to exit on failure for github to distinguish failed from passed tests.