Add AnchorProgram test project
The AnchorProgram node is untested and outdated. This commit adds an
initial test structure and a simple test for fetching idl. The workflow is
modified to run the local validator with test accounts initialized by
searching the test_accounts folder for json account files. This commit
also includes a fix for async idl fetching.