Rename keypair get/set private/public value
Value is not good name. String might more accurately describe what the
user can expect from the function. String implies that the return type is
encoded somehow and the user can derrive that it is the bs58 encoded
value. With the name value it can be any value and that is confusing. The
examples are updated to use the new names.