Restructure SolanaClient to allow multiple callbacks
The HttpClient only allows for one response callback. All requests are
affected when changing the callback. This commit allows internal requests
to have distinct callbacks. It allows this by modifying the internal Rpc
client to take a callback in its request queue. This commit also affects
the WS client with a refactor, solving unsubscribe and ws port problems.