Add HoneyComb node and code generator
HoneyComb interface is implemented and a generator for their graphQL
methods is implemented. HoneyComb graphQL API is queried through a new
QL client and the response is compiled into a transaction.

Add ShdwDrive Node

The ShdwDrive can be used to create a storage and upload files to it.
It is implemented by having a client query their API to pass transactions
that will be transfered by their service.