Add poll signature fallback mechanism for transactions
A fallback mechanism for transactions is implementd. When the web sockets
fail, the transactions need an extra mechanism when this happens to make
sure the confirmation signals are emitted. This mechanism is implemented
by starting a poll mode when web socket packets are not received for a
certain amount of time.