Fix signature verification failure for web wallets
Browser wallets fail whenever they alter the transactions. Since most
wallets tend to modify the transactions more often, this commit
introduces a way to adapt to the changes made by browser wallets.
A method for detecting modifications is implemented. When modifications
are detected, the transaction will be rebuilt from the browser wallet.