Fix crash on set payer on deserialized transaction
When calling set_payer on a deserialized transaction, the signature state
is cleared. This will also cause the transaction message to be recreated.
This doesn't work since the data to create the message is unavailable.
These changes adds a flag that indicates if the transaction is
deserialized. If it is the set_payer method will warn user and do nothing.