Remove errors at startup
The errors related to autoload plugin are caused by godot parsing gd files
when autoloads are disabled. This commit avoids the problem by keeping
SolanaService as an optional utility. It can be toggled from the Solana
menu and kept disabled as default. The toggle mechanism keeps the optional
utilities in hidden folders to avoid gd parsing.